requirements.txt
pandas
langchain
streamlit
transformers
faiss-cpu
tensorflow
openpyxl
tf-keras
fastapi
uvicorn